# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClinicalCompass API",
    "summary": "Authentication API with database-backed opaque tokens.",
    "description": "\n**ClinicalCompass API** \u2014 username/password authentication using opaque,\nserver-side tokens (no JWT).\n\n### Auth flow\n1. `POST /auth/register` to create a user.\n2. `POST /auth/login` with the username + password to receive an opaque\n   `access_token`.\n3. Send the token as `Authorization: Bearer <access_token>` on protected\n   endpoints.\n4. `POST /auth/logout` to revoke the token.\n\nTokens are random strings stored in the database and expire after 24 hours.\n",
    "contact": {
      "name": "ClinicalCompass",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/auth/username": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Check whether a username is available",
        "description": "Returns whether the given username is free. Use this to validate the username a user typed before registering.",
        "operationId": "username_available_auth_username_get",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "examples": [
                "jane.smith"
              ],
              "title": "Username"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsernameAvailableResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Generate an available username from a full name",
        "description": "Derives a username from the given full name (e.g. 'Dr. Jane Smith' -> 'jane.smith'), checks it against existing users, and returns a free username plus a few alternatives. Creates nothing.",
        "operationId": "generate_username_auth_username_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsernameGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsernameGenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register a new user",
        "description": "Creates a user account from full name, email, password and role. A username may be supplied; if omitted it is generated from the full name (e.g. 'Dr. Jane Smith' -> 'jane.smith', with a numeric suffix on collision). The password is hashed with bcrypt before storage. Username and email must each be unique.",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "409": {
            "description": "Username or email already taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Log in and obtain an access token",
        "description": "Validates the email and password and returns an opaque bearer token (no JWT) that expires after 24 hours.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "User account is disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Request a password-reset code (OTP)",
        "description": "Emails a short-lived, one-time numeric code (OTP) to the account with the given email. The response is identical whether or not the email is registered, to avoid account enumeration. Any previously issued codes for the account are invalidated.\n\nThe code is only ever delivered by email and is never returned in the response. When no mailer is configured (local development) it is written to the server log instead.",
        "operationId": "forgot_password_auth_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Reset the password using an emailed OTP",
        "description": "Sets a new password given the email and the one-time code sent by /auth/forgot-password. On success the code is consumed and all of the user's existing access tokens are revoked, forcing a fresh login. The code is invalidated after too many incorrect attempts.",
        "operationId": "reset_password_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid or expired code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Log out (revoke the current token)",
        "description": "Deletes the supplied bearer token from the database, immediately revoking it.",
        "operationId": "logout_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get the current authenticated user",
        "description": "Returns the user that owns the supplied bearer token.",
        "operationId": "me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "User account is disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Invalid username or password"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Standard FastAPI error envelope."
      },
      "ForgotPasswordRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "drsmith@example.com"
            ]
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest"
      },
      "ForgotPasswordResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Generic confirmation, the same regardless of whether the email exists, to avoid account enumeration."
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At",
            "description": "When the code expires (null when no code was issued)."
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ForgotPasswordResponse",
        "example": {
          "expires_at": "2026-06-08T11:42:05.227328Z",
          "message": "If the email is registered, a reset code has been sent."
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "drsmith@example.com"
            ]
          },
          "password": {
            "type": "string",
            "title": "Password",
            "examples": [
              "s3cret-pass"
            ]
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "TokenOut": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_at"
        ],
        "title": "TokenOut",
        "example": {
          "access_token": "MDJ3cy0P_g5Rk8VD_fnkMAc_VSG2kU1MJRY34Ic52oU",
          "expires_at": "2026-06-02T11:32:05.227328Z",
          "token_type": "bearer"
        }
      },
      "UserCreate": {
        "properties": {
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 3,
                "pattern": "^[a-zA-Z0-9._-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username",
            "description": "Optional. If omitted, one is generated from the full name.",
            "examples": [
              "jane.smith"
            ]
          },
          "full_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Full Name",
            "examples": [
              "Dr. Jane Smith"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "examples": [
              "drsmith@example.com"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "examples": [
              "s3cret-pass"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "examples": [
              "licensed_professional"
            ]
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "email",
          "password",
          "role"
        ],
        "title": "UserCreate"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "full_name",
          "role",
          "email",
          "is_active",
          "created_at"
        ],
        "title": "UserOut",
        "example": {
          "created_at": "2026-06-01T11:32:04.715581Z",
          "email": "drsmith@example.com",
          "full_name": "Dr. Jane Smith",
          "id": 1,
          "is_active": true,
          "role": "licensed_professional",
          "username": "drsmith"
        }
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "medical_student",
          "licensed_professional"
        ],
        "title": "UserRole",
        "description": "Whether the user is a student or a licensed practitioner."
      },
      "UsernameAvailableResponse": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "examples": [
              "jane.smith"
            ]
          },
          "available": {
            "type": "boolean",
            "title": "Available",
            "description": "Whether the username is free to use",
            "examples": [
              true
            ]
          }
        },
        "type": "object",
        "required": [
          "username",
          "available"
        ],
        "title": "UsernameAvailableResponse"
      },
      "UsernameGenerateRequest": {
        "properties": {
          "full_name": {
            "type": "string",
            "minLength": 1,
            "title": "Full Name",
            "examples": [
              "Dr. Jane Smith"
            ]
          }
        },
        "type": "object",
        "required": [
          "full_name"
        ],
        "title": "UsernameGenerateRequest"
      },
      "UsernameGenerateResponse": {
        "properties": {
          "full_name": {
            "type": "string",
            "title": "Full Name",
            "examples": [
              "Dr. Jane Smith"
            ]
          },
          "username": {
            "type": "string",
            "title": "Username",
            "description": "A free username derived from the full name",
            "examples": [
              "jane.smith"
            ]
          },
          "alternatives": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Alternatives",
            "examples": [
              [
                "jane.smith1",
                "jane.smith2"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "username"
        ],
        "title": "UsernameGenerateResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}